Move `GetHostDistroName()` to `distro` package
This will allow to use the function also in `osbuild-composer`, while
maintaining only a single copy of the code. I chose this approach
instead of making the whole `internal/common` package public. The
function is related to distros, so this move also makes some sense.
